Change JwtTokenService create select for update with Lock and delete. Return an exception when refresh token is not found

+++ modules/authorization/src/main/java/it/smartcommunitylabdhub/authorization/services/JwtTokenService.java
@@ -25,11 +25,14 @@ import it.smartcommunitylabdhub.authorization.repositories.RefreshTokenRepositor
 import it.smartcommunitylabdhub.authorization.utils.JWKUtils;
 import it.smartcommunitylabdhub.commons.config.ApplicationProperties;
 import it.smartcommunitylabdhub.commons.config.SecurityProperties;
+
 import java.text.ParseException;
 import java.time.Instant;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
+
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -132,13 +135,13 @@ public class JwtTokenService implements InitializingBean {
         SignedJWT refreshToken = generateRefreshToken(authentication, accessToken);
 
         return TokenResponse
-            .builder()
-            .accessToken(accessToken)
-            .refreshToken(refreshToken)
-            .expiration(accessTokenDuration)
-            .clientId(clientId)
-            .issuer(applicationProperties.getEndpoint())
-            .build();
+                .builder()
+                .accessToken(accessToken)
+                .refreshToken(refreshToken)
+                .expiration(accessTokenDuration)
+                .clientId(clientId)
+                .issuer(applicationProperties.getEndpoint())
+                .build();
     }
 
     public String generateAccessTokenAsString(Authentication authentication) throws JwtTokenServiceException {
@@ -165,19 +168,19 @@ public class JwtTokenService implements InitializingBean {
 
             // build access token claims
             JWTClaimsSet.Builder claims = new JWTClaimsSet.Builder()
-                .subject(authentication.getName())
-                .issuer(applicationProperties.getEndpoint())
-                .issueTime(Date.from(now))
-                .audience(applicationProperties.getName())
-                .jwtID(UUID.randomUUID().toString())
-                .expirationTime(Date.from(now.plusSeconds(accessTokenDuration)));
+                    .subject(authentication.getName())
+                    .issuer(applicationProperties.getEndpoint())
+                    .issueTime(Date.from(now))
+                    .audience(applicationProperties.getName())
+                    .jwtID(UUID.randomUUID().toString())
+                    .expirationTime(Date.from(now.plusSeconds(accessTokenDuration)));
 
             //define authorities as claims
             List<String> authorities = authentication
-                .getAuthorities()
-                .stream()
-                .map(GrantedAuthority::getAuthority)
-                .toList();
+                    .getAuthorities()
+                    .stream()
+                    .map(GrantedAuthority::getAuthority)
+                    .toList();
 
             claims.claim("authorities", authorities);
 
@@ -200,7 +203,7 @@ public class JwtTokenService implements InitializingBean {
     }
 
     public SignedJWT generateRefreshToken(Authentication authentication, SignedJWT accessToken)
-        throws JwtTokenServiceException {
+            throws JwtTokenServiceException {
         if (signer == null) {
             throw new UnsupportedOperationException("signer not available");
         }
@@ -212,12 +215,12 @@ public class JwtTokenService implements InitializingBean {
 
             // build refresh token claims
             JWTClaimsSet.Builder claims = new JWTClaimsSet.Builder()
-                .subject(authentication.getName())
-                .issuer(applicationProperties.getEndpoint())
-                .issueTime(Date.from(now))
-                .audience(applicationProperties.getName())
-                .jwtID(UUID.randomUUID().toString())
-                .expirationTime(Date.from(now.plusSeconds(refreshTokenDuration)));
+                    .subject(authentication.getName())
+                    .issuer(applicationProperties.getEndpoint())
+                    .issueTime(Date.from(now))
+                    .audience(applicationProperties.getName())
+                    .jwtID(UUID.randomUUID().toString())
+                    .expirationTime(Date.from(now.plusSeconds(refreshTokenDuration)));
 
             //associate access token via hash binding
             String hash = JWKUtils.getAccessTokenHash(jwsAlgorithm, accessToken);
@@ -225,10 +228,10 @@ public class JwtTokenService implements InitializingBean {
 
             //define authorities as claims
             List<String> authorities = authentication
-                .getAuthorities()
-                .stream()
-                .map(GrantedAuthority::getAuthority)
-                .toList();
+                    .getAuthorities()
+                    .stream()
+                    .map(GrantedAuthority::getAuthority)
+                    .toList();
 
             claims.claim("authorities", authorities);
 
@@ -245,14 +248,14 @@ public class JwtTokenService implements InitializingBean {
 
             // Store Refresh Token into db.
             refreshTokenRepository.save(
-                RefreshTokenEntity
-                    .builder()
-                    .id(claimsSet.getJWTID())
-                    .subject(authentication.getName())
-                    .refreshToken(jwt.serialize())
-                    .issuedTime(claimsSet.getIssueTime())
-                    .expirationTime(claimsSet.getExpirationTime())
-                    .build()
+                    RefreshTokenEntity
+                            .builder()
+                            .id(claimsSet.getJWTID())
+                            .subject(authentication.getName())
+                            .refreshToken(jwt.serialize())
+                            .issuedTime(claimsSet.getIssueTime())
+                            .expirationTime(claimsSet.getExpirationTime())
+                            .build()
             );
 
             return jwt;
@@ -292,7 +295,14 @@ public class JwtTokenService implements InitializingBean {
                 throw new JwtTokenServiceException("Token subject does not match authentication subject");
             }
 
-            // Remove the token from the repository
+            // Lock the token
+            Optional<RefreshTokenEntity> tokenEntityOptional = refreshTokenRepository.findByIdForUpdate(tokenId);
+
+            if (tokenEntityOptional.isEmpty()) {
+                throw new JwtTokenServiceException("Refresh token does not exist");
+            }
+
+            // Delete the token
             refreshTokenRepository.deleteById(tokenId);
 
             log.info("Refresh token successfully consumed and removed from repository");
